Refactored Page folder structure
