--- src/pages/loginuser/index.js
@@ -1,3 +1,0 @@
-import LoginUser from "./loginuser"
-
-export default LoginUser
