--- src/pages/registeruser/index.js
@@ -1,3 +1,0 @@
-import RegisterUser from "./registeruser"
-
-export default RegisterUser
