+++ src/pages/sketch/view/index.js
@@ -1,0 +1,3 @@
+import SketchView from "./sketchview"
+
+export default SketchView
