+++ src/pages/static/terms/index.js
