+++ src/pages/static/terms/terms.js
