+++ src/pages/static/terms/terms.scss
