+++ src/pages/user/login/index.js
@@ -1,0 +1,3 @@
+import UserLogin from "./userlogin"
+
+export default UserLogin
