+++ src/pages/user/register/index.js
@@ -1,0 +1,3 @@
+import UserRegister from "./userregister"
+
+export default UserRegister
