+++ src/pages/user/register/userregister.scss
@@ -1,3 +1,3 @@
-.loginuser {
+.userregister {
     margin: 4em auto;
 }
