+++ src/pages/user/view/index.js
@@ -1,0 +1,3 @@
+import UserView from "./userview"
+
+export default UserView
