--- src/pages/viewproject/index.js
@@ -1,3 +1,0 @@
-import ViewProject from "./viewproject"
-
-export default ViewProject
