--- src/pages/viewuser/index.js
@@ -1,3 +1,0 @@
-import ViewUser from "./viewuser"
-
-export default ViewUser
