+++ src/routes.js
@@ -1,24 +1,24 @@
 import Homepage from "./pages/homepage"
-import RegisterUser from "./pages/registeruser"
-import LoginUser from "./pages/loginuser"
-import Terms from "./pages/terms"
-import ViewUser from "./pages/viewuser"
-import ViewProject from "./pages/viewproject"
-import ViewPolyProject from "./pages/poly/viewproject"
-import ViewPolyAll from "./pages/poly/viewall"
+import UserRegister from "./pages/user/login"
+import UserLogin from "./pages/user/register"
+import UserView from "./pages/user/view"
+import Terms from "./pages/static/terms"
+import SketchView from "./pages/sketch/view"
+import PolySketchView from "./pages/poly/sketchview"
+import PolyViewAll from "./pages/poly/viewall"
 import Dashboard from "./pages/dashboard"
-import NotFound from "./pages/NotFound"
+import NotFound from "./pages/notfound"
 
 export const Routes = [
   {
     path: "/register",
     exact: true,
-    component: RegisterUser,
+    component: UserRegister,
   },
   {
     path: "/login",
     exact: true,
-    component: LoginUser,
+    component: UserLogin,
   },
   {
     path: "/terms",
@@ -34,22 +34,22 @@ export const Routes = [
   {
     path: "/user/:id",
     exact: true,
-    component: ViewUser,
+    component: UserView,
   },
   {
     path: "/view/:id",
     exact: true,
-    component: ViewProject,
+    component: SketchView,
   },
   {
     path: "/poly",
     exact: true,
-    component: ViewPolyAll,
+    component: PolyViewAll,
   },
   {
     path: "/poly/view/:id",
     exact: true,
-    component: ViewPolyProject,
+    component: PolySketchView,
   },
   {
     path: "/",
